+++ src/main/java/kr/hhplus/be/server/api/concert/presentation/controller/ConcertController.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.api.concert.presentation;
+package kr.hhplus.be.server.api.concert.presentation.controller;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -7,7 +7,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import kr.hhplus.be.server.common.response.ApiResponse;
-import kr.hhplus.be.server.api.concert.domain.service.ConcertService;
+import kr.hhplus.be.server.api.concert.application.service.ConcertService;
 import kr.hhplus.be.server.api.concert.application.dto.ConcertScheduleResult;
 import kr.hhplus.be.server.api.concert.application.dto.ConcertSeatResult;
 import kr.hhplus.be.server.api.concert.presentation.dto.ConcertScheduleResponse;
