+++ src/main/java/kr/hhplus/be/server/api/user/application/service/BalanceHistoryService.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.api.user.domain.service;
+package kr.hhplus.be.server.api.user.application.service;
 
 import kr.hhplus.be.server.api.user.application.dto.BalanceHistoryResult;
 import kr.hhplus.be.server.api.user.application.dto.BalanceHistoryDto;
